Queue DELETE instructions for virtual nodes
Summary:
To ensure that we're eventually deleting all ShadowViews from the StubViewTree and from the mounting layer, make sure that we always queue conflicts if they're DELETE instructions, including for virtual nodes.

I was able to hit this by running some extremely complex animations. It is extremely unlikely that even a single user is hitting this in prod. Therefore, while this is the right change to make, I don't expect (for example) OOMs to go down at all.

Changelog: [Internal]

Reviewed By: mdvacca

Differential Revision: D27492935

fbshipit-source-id: bce332feb15229af271cc6e14b8367ebcb36536b